fix(frontend): use reusable DashboardTimelineItem for upcoming and activity sections
